refactor(search): Regroup all components in the same folder
