+++ src/modules/search/components/AppBarSearch.spec.jsx
@@ -1,17 +1,17 @@
-import { render } from '@testing-library/react'
+import { render, screen } from '@testing-library/react'
 import React from 'react'
 
 import CozyClient from 'cozy-client'
 
-import AppBarSearch from 'modules/search/AppBarSearch'
+import AppBarSearch from 'modules/search/components/AppBarSearch'
 import AppLike from 'test/components/AppLike'
 
 it('should display the Searchbar', () => {
   const client = new CozyClient({})
-  const appBarSearch = render(
+  render(
     <AppLike client={client}>
       <AppBarSearch />
     </AppLike>
   )
-  expect(appBarSearch).toMatchSnapshot()
+  expect(screen.getByPlaceholderText('Search anything')).toBeInTheDocument()
 })
